# apis/restaurant/src/swagger 2.yaml
swagger: '2.0'
info:
  version: 1.0.0
  title: Restaurants micro-service API Doc
  description: API for managing restaurants
servers:
  - url: http://localhost:3006
basePath: /
schemes:
  - http
consumes:
  - application/json
produces:
  - application/json
definitions:
  Rate:
    type: object
    properties:
      id_user:
        type: integer
      rate:
        type: integer
      comment:
        type: string
      date:
        type: string
        format: date-time
  Item:
    type: object
    properties:
      label:
        type: string
      description:
        type: string
      photo:
        type: string
      price:
        type: number
      type:
        type: string
        enum: ['plat', 'boisson', 'sauce', 'accompagnement']
  MenuItem:
    type: object
    properties:
      item:
        $ref: '#/definitions/Item'
      quantity:
        type: integer
  Menu:
    type: object
    properties:
      label:
        type: string
      description:
        type: string
      photo:
        type: string
      price:
        type: number
      items:
        type: array
        items:
          $ref: '#/definitions/MenuItem'
  Restaurant:
    type: object
    properties:
      id_credentials:
        type: string
      name:
        type: string
      photo:
        type: string
      schedule:
        type: string
      id_address:
        type: string
      category:
        type: string
      mean_rate:
        type: number
      rates:
        type: array
        items:
          $ref: '#/definitions/Rate'
      menus:
        type: array
        items:
          $ref: '#/definitions/Menu'
      items:
        type: array
        items:
          $ref: '#/definitions/Item'
paths:
  /restaurants:
    post:
      tags:
        - Restaurants
      summary: 'Create a new restaurant'
      operationId: 'createRestaurant'
      parameters:
        - in: body
          name: body
          required: true
          schema:
            $ref: '#/definitions/Restaurant'
      responses:
        '201':
          description: 'Restaurant created successfully'
        '500':
          description: 'Server error'
    get:
      tags:
        - Restaurants
      summary: 'Get a specific restaurant'
      operationId: 'getRestaurant'
      parameters:
        - in: path
          name: restaurantId
          required: true
          type: string
      responses:
        '200':
          description: 'Successful operation'
          schema:
            $ref: '#/definitions/Restaurant'
        '404':
          description: 'Restaurant not found'
        '500':
          description: 'Server error'
  /restaurants-categories:
    get:
      tags:
        - Restaurants
      summary: 'Get all restaurant categories'
      operationId: 'getRestaurantCategories'
      responses:
        '200':
          description: 'Successful operation'
        '500':
          description: 'Server error'
  /restaurants-top-rated:
    get:
      tags:
        - Restaurants
      summary: 'Get top 10 rated restaurants'
      operationId: 'getTopRatedRestaurants'
      responses:
        '200':
          description: 'Successful operation'
        '500':
          description: 'Server error'
...